Adding links to data formats
As someone that found Jet via a technology news feed, I was curious.
I did not know what about the EDN nor Transit format. With this commit
my intention is to provide a bit of wayfinding documentation for those
that are also curious about Jet and it's related data formats.

Closes borkdude/jet#63