Added POSIX style flags, with single letters
Added them because the flags with double hyphens followed by a word,
e.g.

    --foo

are a GNU-ism. Whereas the single letter flags are POSIX
recommendation, see:
https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap12.html

That is the official blurb, the real motivation was that since this is
a CLI tool, it would be much quicker to type single letter flags than
the double hyphen plus a word flags.